CDF_CSE/requirements.txt
numpy==1.26.4
pandas==2.2.2
PyYAML==6.0.1
matplotlib==3.8.4
seaborn==0.13.2
tqdm==4.66.4
pytest==8.2.2
